move `function.py` under `fpy2`
