+++ fpy2/ast/__init__.py
@@ -7,6 +7,7 @@ from .formatter import Formatter
 from .fpyast import set_default_formatter
 from .visitor import AstVisitor, DefaultAstVisitor, DefaultAstTransformVisitor
 
+from .context_inline import ContextInline
 from .define_use import DefineUse
 from .live_vars import LiveVars
 from .syntax_check import SyntaxCheck
