+++ fpy2/runtime/__init__.py
@@ -1,6 +1,1 @@
 from .env import ForeignEnv
-from .function import (
-    Function,
-    get_default_function_call,
-    set_default_function_call,
-)
